backend/requirements.txt
# D:\kodingan\backend\requirements.txt
fastapi
uvicorn
Pillow
python-multipart
gunicorn # <--- TAMBAH INI